Dev jcy

+++ pizza-data/src/main/java/edu.ecnu.scsse.pizza.data/repository/AdminJpaRepository.java
@@ -3,6 +3,6 @@ package edu.ecnu.scsse.pizza.data.repository;
 import edu.ecnu.scsse.pizza.data.domain.AdminEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminJpaRepository extends JpaRepository<AdminEntity,Long> {
+public interface AdminJpaRepository extends JpaRepository<AdminEntity,Integer> {
 
 }
