+++ pizza-data/src/main/java/edu.ecnu.scsse.pizza.data/repository/CommentJpaRepository.java
@@ -3,6 +3,6 @@ package edu.ecnu.scsse.pizza.data.repository;
 import edu.ecnu.scsse.pizza.data.domain.CommentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentJpaRepository extends JpaRepository<CommentEntity,Long> {
+public interface CommentJpaRepository extends JpaRepository<CommentEntity,Integer> {
 
 }
