+++ pizza-data/src/main/java/edu.ecnu.scsse.pizza.data/repository/ShopIngredientJpaRepository.java
@@ -3,6 +3,6 @@ package edu.ecnu.scsse.pizza.data.repository;
 import edu.ecnu.scsse.pizza.data.domain.ShopIngredientEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShopIngredientJpaRepository extends JpaRepository<ShopIngredientEntity,Long> {
+public interface ShopIngredientJpaRepository extends JpaRepository<ShopIngredientEntity,Integer> {
 
 }
